* more refactoring * averaging error rate
